pdf-docx-zip-python/app/requirements.txt
pymupdf==1.24.0
flask==3.0.3
python-docx==1.1.0
python-magic==0.4.27
libmagic
gunicorn==20.1.0
Flask-Bootstrap==3.3.7.1